chore(rails plugin): make it look like a rails plugin
### What

This is a B.A.C. (big-ass-commit) that gets this gem looking the same as
if you'd run `rails plugin new pluck_to_struct`.

### Why

Making everything look and feel familiar for myself and contributors is
a win. Also, we can follow the conventions of a "dummy" app and avoid
hand-rolled database migrations, seeding, etc. in the tests.